backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pandas>=2.3.0
scikit-learn>=1.7.0
joblib>=1.4.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
